pyproject.toml
[project]
name = "desafio-senior-backend-developer"
version = "1.0.0"
description = ""
authors = [
    {name = "Jean Torre",email = "jean.torre21@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-material (>=9.6.12,<10.0.0)",
    "mkdocstrings-python (>=1.16.10,<2.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "ignr (>=2.2,<3.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uuid (>=1.30,<2.0)",
    "pytz (>=2025.2,<2026.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "alembic (>=1.15.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = 'black'
line_length = 89
skip_gitignore = true

[tool.black]
line-length = 89
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | _build
  | build
  | dist
)/
'''

[tool.taskipy.tasks]
tests = "PYTHONPATH=./backend pytest -v"
docs = "mkdocs serve"
deploy = "mkdocs gh-deploy"
migrate = "alembic -c alembic.ini upgrade head"
makemigrations = "alembic -c alembic.ini revision --autogenerate -m"
downgrade = "alembic -c alembic.ini downgrade -1"
